gena_web/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "gena"
version = "0.0.1"
description = "GenA"
authors = [
  { name = "RANEPA" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "loguru",
    "python-dotenv",
    "tqdm",
    "chonkie (>=1.1.0,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "streamlit (>=1.46.1,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "plotly (>=5.0.0,<6.0.0)",
    "matplotlib",
    "extra-streamlit-components (>=0.1.70,<1.0.0)"
]
requires-python = "~=3.12.0"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.isort]
profile = "black"
known_first_party = ["gena"]
force_sort_within_sections = true

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
